backend/requirement.txt
Web Framework

Flask==3.0.2
Flask-Cors==4.0.0
gunicorn==21.2.0

Database & Auth

firebase-admin==6.4.0

Cryptography (Matches your crypto_vault.py)

pycryptodome==3.20.0

Image & Math Processing

Pillow==10.2.0
numpy==1.26.4

Note: FFmpeg must be installed on the system (OS level), not via pip.